refactor: DDD 구조로 리팩터링
