--- src/main/java/org/sopt/controller/UserController.java
@@ -1,27 +1,0 @@
-package org.sopt.controller;
-
-import org.sopt.dto.request.UserCreateRequest;
-import org.sopt.dto.response.SuccessResponse;
-import org.sopt.service.UserService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/users")
-public class UserController {
-
-    private final UserService userService;
-
-    public UserController(UserService userService){
-        this.userService = userService;
-    }
-
-    @PostMapping
-    public ResponseEntity<SuccessResponse<String>> createUser(@RequestBody UserCreateRequest userCreateRequest){
-        userService.saveUser(userCreateRequest);
-        return ResponseEntity.ok(new SuccessResponse<>(ResponseMessage.CREATE_USER_SUCCESS.getMessage()));
-    }
-}
