--- src/main/java/org/sopt/domain/Comment.java
@@ -1,33 +1,0 @@
-package org.sopt.domain;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Size;
-import lombok.*;
-
-@Entity
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Comment extends BaseTimeEntity{
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(columnDefinition = "TEXT")
-    @Size(min = 1, max = 300)
-    private String comment;
-
-    @ManyToOne
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    public void updateComment(String comment){
-        this.comment = comment;
-    }
-}
