--- src/main/java/org/sopt/domain/User.java
@@ -1,37 +1,0 @@
-package org.sopt.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private String email;
-
-    private String password;
-
-    @OneToMany(mappedBy = "user", fetch = FetchType.LAZY)
-    private List<Post> posts;
-
-    @OneToMany(mappedBy = "user")
-    private List<Comment> comments;
-
-    public User(String name, String email){
-        this.name = name;
-        this.email = email;
-    }
-
-}
