+++ src/main/java/org/sopt/domain/post/repository/CommentRepository.java
@@ -1,0 +1,16 @@
+package org.sopt.domain.post.repository;
+
+import org.sopt.domain.post.model.PostComment;
+import org.sopt.domain.post.model.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface CommentRepository extends JpaRepository<PostComment, Long> {
+
+    List<PostComment> findByPost(Post post);
+    Optional<PostComment> findByIdAndUserIdAndPostId(Long id, Long userId, Long postId);
+}
