+++ src/main/java/org/sopt/domain/user/model/User.java
@@ -1,0 +1,38 @@
+package org.sopt.domain.user.model;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.sopt.domain.post.model.PostComment;
+import org.sopt.domain.post.model.Post;
+
+import java.util.List;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    private String email;
+
+    private String password;
+
+    @OneToMany(mappedBy = "user", fetch = FetchType.LAZY)
+    private List<Post> posts;
+
+    @OneToMany(mappedBy = "user")
+    private List<PostComment> postComments;
+
+    public User(String name, String email){
+        this.name = name;
+        this.email = email;
+    }
+
+}
