+++ src/main/java/org/sopt/domain/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package org.sopt.domain.user.repository;
+
+import org.sopt.domain.user.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
