--- src/main/java/org/sopt/dto/response/ErrorResponse.java
@@ -1,14 +1,0 @@
-package org.sopt.dto.response;
-
-import org.sopt.exception.ErrorCode;
-
-public record ErrorResponse(
-        Boolean success,
-        int status,
-        String code,
-        String message
-) {
-    public static ErrorResponse from(ErrorCode errorCode){
-        return new ErrorResponse(false, errorCode.getStatus(), errorCode.getCode(), errorCode.getMessage());
-    }
-}
