--- src/main/java/org/sopt/dto/response/PostDetailResponseDto.java
@@ -1,18 +1,0 @@
-package org.sopt.dto.response;
-
-import org.sopt.domain.Post;
-
-import java.time.LocalDateTime;
-
-public record PostDetailResponseDto(
-        Long id,
-        String title,
-        String content,
-        String userName,
-        LocalDateTime createdAt,
-        LocalDateTime modifiedAt
-) {
-    public static PostDetailResponseDto from(Post post){
-        return new PostDetailResponseDto(post.getId(), post.getTitle(), post.getContent(), post.getUser().getName(), post.getCreatedDate(), post.getModifiedDate());
-    }
-}
