--- src/main/java/org/sopt/dto/response/SuccessResponse.java
@@ -1,7 +1,0 @@
-package org.sopt.dto.response;
-
-public record SuccessResponse<T>(Boolean success, T data) {
-    public SuccessResponse(T data){
-        this(true, data);
-    }
-}
