--- src/main/java/org/sopt/exception/CustomAccessDeniedException.java
@@ -1,16 +1,0 @@
-package org.sopt.exception;
-
-public class CustomAccessDeniedException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public CustomAccessDeniedException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode(){
-        return this.errorCode;
-    }
-}
-
