--- src/main/java/org/sopt/exception/CustomNotFoundException.java
@@ -1,14 +1,0 @@
-package org.sopt.exception;
-
-public class CustomNotFoundException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public CustomNotFoundException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-}
