+++ src/main/java/org/sopt/global/dto/SuccessResponse.java
@@ -1,0 +1,7 @@
+package org.sopt.global.dto;
+
+public record SuccessResponse<T>(Boolean success, T data) {
+    public SuccessResponse(T data){
+        this(true, data);
+    }
+}
