+++ src/main/java/org/sopt/global/exception/CustomAccessDeniedException.java
@@ -1,0 +1,18 @@
+package org.sopt.global.exception;
+
+import org.sopt.global.ErrorCode;
+
+public class CustomAccessDeniedException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public CustomAccessDeniedException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+    public ErrorCode getErrorCode(){
+        return this.errorCode;
+    }
+}
+
