+++ src/main/java/org/sopt/global/exception/CustomNotFoundException.java
@@ -1,0 +1,16 @@
+package org.sopt.global.exception;
+
+import org.sopt.global.ErrorCode;
+
+public class CustomNotFoundException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public CustomNotFoundException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+    public ErrorCode getErrorCode() {
+        return errorCode;
+    }
+}
