--- src/main/java/org/sopt/repository/CommentRepository.java
@@ -1,16 +1,0 @@
-package org.sopt.repository;
-
-import org.sopt.domain.Comment;
-import org.sopt.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-
-    List<Comment> findByPost(Post post);
-    Optional<Comment> findByIdAndUserIdAndPostId(Long id, Long userId, Long postId);
-}
