--- src/main/java/org/sopt/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package org.sopt.repository;
-
-import org.sopt.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
